fix: android options numbers as double

+++ android/src/main/java/com/reactnativegps/LocationUpdates.kt
@@ -183,10 +183,10 @@ class LocationUpdates(private val context: Context): ServiceLifecycle, ServiceIn
         val androidOptions = options?.get("android") as HashMap<String, Any>?
 
         val locationOptions = androidOptions?.get("location") as HashMap<String, Any>?
-        val newLocationPriority = locationOptions?.get("priority") as Int?;
+        val newLocationPriority = locationOptions?.get("priority") as Double?;
 
         if (newLocationPriority != null) {
-            locationPriority = newLocationPriority
+            locationPriority = newLocationPriority.toInt()
             val locationRequest = createLocationRequest()
             if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
                 mLocationUpdatesTask = mFusedLocationClient?.requestLocationUpdates(locationRequest, mLocationCallback, Looper.getMainLooper())
